Self-Signed SSL configurations
Adding handling for self-signed SSL configurations, as well as the proxy hard-restart bug when the remote pool can't be contacted in time.  Pools on the master now start in a disabled state until they can be connected to.

If they are shunted into a reconnecting state, the system tags them as offline.

This should also fix the failover systems.  Pools that go inactive due to disconnection will have a disablePool notice sent w/ the pool's hostname.  As soon as the pool comes back (Connect succeeds and login happens), the system will reactivate them.